Multi open source and change to long timestamp and tid

+++ tspoon/src/main/java/it/polimi/affetti/tspoon/evaluation/SimulateRecovery.java
@@ -34,7 +34,7 @@ public class SimulateRecovery {
             System.out.println(updates);
         }
 
-        int wm = wal.getSnapshotInProgressWatermark();
+        long wm = wal.getSnapshotInProgressWatermark();
 
         System.out.println();
         System.out.println(">>> WM:\t\t" + wm);
