+++ tspoon/src/main/java/it/polimi/affetti/tspoon/tgraph/TransactionResult.java
@@ -5,11 +5,11 @@ import org.apache.flink.api.java.tuple.Tuple5;
 /**
  * Created by affo on 26/07/17.
  */
-public class TransactionResult extends Tuple5<Integer, Integer, Object, Vote, Updates> {
+public class TransactionResult extends Tuple5<Long, Long, Object, Vote, Updates> {
     public TransactionResult() {
     }
 
-    public TransactionResult(Integer tid, Integer timestamp, Object originalRecord, Vote vote, Updates updates) {
+    public TransactionResult(Long tid, Long timestamp, Object originalRecord, Vote vote, Updates updates) {
         super(tid, timestamp, originalRecord, vote, updates);
     }
 }
