+++ tspoon/src/main/java/it/polimi/affetti/tspoon/tgraph/state/PL2Strategy.java
@@ -10,7 +10,7 @@ import it.polimi.affetti.tspoon.tgraph.db.ObjectVersion;
 public class PL2Strategy extends PL1Strategy {
 
     @Override
-    public <V> ObjectHandler<V> readVersion(int tid, int timestamp, int watermark, Object<V> versions) {
+    public <V> ObjectHandler<V> readVersion(long tid, long timestamp, long watermark, Object<V> versions) {
         return versions.readCommittedBefore(watermark);
     }
 }
