+++ tspoon/src/main/java/it/polimi/affetti/tspoon/tgraph/state/PL3Strategy.java
@@ -7,8 +7,8 @@ import it.polimi.affetti.tspoon.tgraph.db.Object;
  */
 public class PL3Strategy extends PL2Strategy {
     @Override
-    public boolean canWrite(int tid, int timestamp, int watermark, Object<?> object) {
-        int lastVersion = object.getLastAvailableVersion().version;
+    public boolean canWrite(long tid, long timestamp, long watermark, Object<?> object) {
+        long lastVersion = object.getLastAvailableVersion().version;
         return watermark >= lastVersion;
     }
 }
